Removed ParticleEmitterSlot
