+++ engine/src/main/java/org/destinationsol/game/ship/hulls/HullConfig.java
@@ -19,7 +19,7 @@ import com.badlogic.gdx.graphics.g2d.TextureAtlas;
 import com.badlogic.gdx.math.Vector2;
 import org.destinationsol.common.Immutable;
 import org.destinationsol.game.item.Engine;
-import org.destinationsol.game.particle.ParticleEmitterSlot;
+import org.destinationsol.game.particle.ParticleEmitter;
 import org.destinationsol.game.ship.AbilityConfig;
 
 import java.util.ArrayList;
@@ -75,16 +75,16 @@ public final class HullConfig {
         return new ArrayList<>(data.gunSlots);
     }
 
-    public ParticleEmitterSlot getParticleEmitterSlot(int slotNr) {
-        return data.particleEmitterSlots.get(slotNr);
+    public ParticleEmitter getParticleEmitter(int slotNr) {
+        return data.particleEmitters.get(slotNr);
     }
 
-    public int getNrOfParticleEmitterSlots() {
-        return data.particleEmitterSlots.size();
+    public int getNrOfParticleEmitters() {
+        return data.particleEmitters.size();
     }
 
-    public List<ParticleEmitterSlot> getParticleEmitterSlotList() {
-        return new ArrayList<>(data.particleEmitterSlots);
+    public List<ParticleEmitter> getParticleEmitterList() {
+        return new ArrayList<>(data.particleEmitters);
     }
 
     public List<Vector2> getLightSourcePositions() {
@@ -181,7 +181,7 @@ public final class HullConfig {
         public Vector2 e1Pos;
         public Vector2 e2Pos;
         public List<GunSlot> gunSlots = new ArrayList<>();
-        public List<ParticleEmitterSlot> particleEmitterSlots = new ArrayList<>();
+        public List<ParticleEmitter> particleEmitters = new ArrayList<>();
         public List<Vector2> lightSrcPoss = new ArrayList<>();
         public float durability;
         public boolean hasBase;
@@ -228,7 +228,7 @@ public final class HullConfig {
             this.origin = new Vector2(src.origin);
             this.shipBuilderOrigin = new Vector2(src.shipBuilderOrigin);
             this.gunSlots.addAll(src.gunSlots);
-            this.particleEmitterSlots.addAll(src.particleEmitterSlots);
+            this.particleEmitters.addAll(src.particleEmitters);
         }
     }
 }
